LPS-54637 split logic into Portal and Indexer subclasses that extend common parent class
