+++ portal-impl/src/com/liferay/portal/search/backgroundtask/ReindexBackgroundTaskExecutor.java
@@ -1,0 +1,54 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.search.backgroundtask;
+
+import com.liferay.portal.kernel.backgroundtask.BackgroundTask;
+import com.liferay.portal.kernel.backgroundtask.BackgroundTaskConstants;
+import com.liferay.portal.kernel.backgroundtask.BackgroundTaskResult;
+import com.liferay.portal.kernel.backgroundtask.BaseBackgroundTaskExecutor;
+
+import java.io.Serializable;
+
+import java.util.Map;
+
+/**
+ * @author Andrew Betts
+ */
+public abstract class ReindexBackgroundTaskExecutor
+	extends BaseBackgroundTaskExecutor {
+
+	public ReindexBackgroundTaskExecutor() {
+		setIsolationLevel(BackgroundTaskConstants.ISOLATION_LEVEL_COMPANY);
+	}
+
+	@Override
+	public BackgroundTaskResult execute(BackgroundTask backgroundTask)
+		throws Exception {
+
+		Map<String, Serializable> taskContextMap =
+			backgroundTask.getTaskContextMap();
+
+		String className = (String)taskContextMap.get("className");
+		long[] companyIds = (long[])taskContextMap.get("companyIds");
+
+		reindex(className, companyIds);
+
+		return BackgroundTaskResult.SUCCESS;
+	}
+
+	protected abstract void reindex(String className, long[] companyIds)
+		throws Exception;
+
+}
