+++ portal-impl/src/com/liferay/portal/search/backgroundtask/ReindexPortalBackgroundTaskExecutor.java
@@ -1,0 +1,47 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.search.backgroundtask;
+
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.search.SearchEngineInitializer;
+
+/**
+ * @author Andrew Betts
+ */
+public class ReindexPortalBackgroundTaskExecutor
+	extends ReindexBackgroundTaskExecutor {
+
+	@Override
+	protected void reindex(String className, long[] companyIds)
+		throws Exception {
+
+		for (long companyId : companyIds) {
+			try {
+				SearchEngineInitializer searchEngineInitializer =
+					new SearchEngineInitializer(companyId);
+
+				searchEngineInitializer.reindex();
+			}
+			catch (Exception e) {
+				_log.error(e, e);
+			}
+		}
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		ReindexPortalBackgroundTaskExecutor.class);
+
+}
